chore: add md5 & sha1
